+++ src/languages/php-template.js
@@ -1,0 +1,21 @@
+/*
+Language: PHP Template
+Requires: xml.js, php.js
+Author: Josh Goebel <hello@joshgoebel.com>
+Website: https://www.php.net
+Category: common
+*/
+
+export default function(hljs) {
+  return {
+    name: "PHP template",
+    subLanguage: 'xml',
+    contains: [
+      {
+        begin: /<\?(php|=)?/,
+        end: /\?>/,
+        subLanguage: 'php',
+      }
+    ]
+  };
+}