+++ src/languages/xml.js
@@ -97,19 +97,6 @@ export default function(hljs) {
         begin: /<\?xml/, end: /\?>/, relevance: 10
       },
       {
-        begin: /<\?(php)?/, end: /\?>/,
-        subLanguage: 'php',
-        contains: [
-          // We don't want the php closing tag ?> to close the PHP block when
-          // inside any of the following blocks:
-          {begin: '/\\*', end: '\\*/', skip: true},
-          {begin: 'b"', end: '"', skip: true},
-          {begin: 'b\'', end: '\'', skip: true},
-          hljs.inherit(hljs.APOS_STRING_MODE, {illegal: null, className: null, contains: null, skip: true}),
-          hljs.inherit(hljs.QUOTE_STRING_MODE, {illegal: null, className: null, contains: null, skip: true})
-        ]
-      },
-      {
         className: 'tag',
         /*
         The lookahead pattern (?=...) ensures that 'begin' only matches